Input: Rip out the old input system and get the other thread running
