+++ Furball.Engine/Engine/Helpers/Profiler.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using Kettu;
@@ -16,7 +17,7 @@ public static class Profiler {
     public static void EndProfileAndPrint(string name) {
         double time = EndProfile(name);
 
-        Kettu.Logger.Log($"Profile {name} took {time}ms!", LoggerLevelProfiler.Instance);
+        System.Diagnostics.Debug.WriteLine($"Profile {name} took {time}ms!");
     }
 
     public static double EndProfile(string name) => ((double)Stopwatch.GetTimestamp() / Stopwatch.Frequency - Profiles[name]) * 1000d;
