+++ Furball.Engine/Engine/Input/Events/CharInputEvent.cs
@@ -1,4 +1,5 @@
 using System;
+using Furball.Engine.Engine.OldInput;
 
 namespace Furball.Engine.Engine.Input.Events;
