+++ Furball.Engine/Engine/Input/Events/KeyEventArgs.cs
@@ -1,4 +1,5 @@
 using System;
+using Furball.Engine.Engine.OldInput;
 using Silk.NET.Input;
 
 namespace Furball.Engine.Engine.Input.Events;
