+++ Furball.Engine/Engine/Input/Events/MouseDragEventArgs.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Numerics;
+using Furball.Engine.Engine.OldInput;
 using Silk.NET.Input;
 
 namespace Furball.Engine.Engine.Input.Events;
