+++ Furball.Engine/Engine/Input/Events/MouseScrollEventArgs.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Numerics;
+using Furball.Engine.Engine.OldInput;
 
 namespace Furball.Engine.Engine.Input.Events;
