+++ Furball.Engine/Engine/Input/Keybind.cs
@@ -1,4 +1,5 @@
 using Furball.Engine.Engine.Config;
+using Furball.Engine.Engine.OldInput;
 using Silk.NET.Input;
 
 namespace Furball.Engine.Engine.Input;
