+++ Furball.Engine/Furball.Engine.csproj
@@ -39,6 +39,7 @@
         <PackageReference Include="JetBrains.Annotations" Version="2022.3.1" />
         <PackageReference Include="Eto.Forms" Version="2.7.2" />
         <PackageReference Include="Eto.Platform.Gtk" Version="2.7.2" />
+        <PackageReference Include="Vanara.PInvoke.Kernel32" Version="3.4.11" />
     </ItemGroup>
 
     <ItemGroup Condition="'$(UseImGui)'=='true'">
