+++ Furball.Engine/FurballGame.cs
@@ -22,10 +22,9 @@ using Furball.Engine.Engine.Graphics.Drawables.Tweens;
 using Furball.Engine.Engine.Graphics.Drawables.Tweens.TweenTypes;
 using Furball.Engine.Engine.Helpers;
 using Furball.Engine.Engine.Helpers.Logger;
-using Furball.Engine.Engine.Input;
-using Furball.Engine.Engine.Input.InputMethods;
 using Furball.Engine.Engine.Localization;
 using Furball.Engine.Engine.Localization.Languages;
+using Furball.Engine.Engine.Input;
 using Furball.Engine.Engine.Platform;
 using Furball.Engine.Engine.Timing;
 using Furball.Engine.Engine.Transitions;
@@ -155,22 +154,23 @@ public class FurballGame : Game {
         _stopwatch.Start();
 
         InputManager = new InputManager();
+        InputManager.Start();
         
-        bool usingEvDevInput = false;
+        // bool usingEvDevInput = false;
 
-        //If we are on linux and we are the root user, we can use the EvDev input instead, which can detect multiple keyboards separately
-        if (RuntimeInfo.CurrentPlatform() == OSPlatform.Linux && UnixEnvironment.IsRoot()) {
-            usingEvDevInput = true;
-            InputManager.RegisterInputMethod(new EvDevKeyboardInputMethod());
-        }
-
-        if (this.WindowManager is SilkWindowManager silkWindowManager) {
-            if(!usingEvDevInput)
-                InputManager.RegisterInputMethod(InputManager.SilkWindowingKeyboardInputMethod = new SilkWindowingKeyboardInputMethod(silkWindowManager.InputContext));
-        
-            InputManager.RegisterInputMethod(InputManager.SilkWindowingMouseInputMethod = new SilkWindowingMouseInputMethod(silkWindowManager.InputContext));
-            InputManager.SilkWindowingMouseInputMethod.SetRawInputStatus(FurballConfig.Instance.RawMouseInput);
-        }
+        // //If we are on linux and we are the root user, we can use the EvDev input instead, which can detect multiple keyboards separately
+        // if (RuntimeInfo.CurrentPlatform() == OSPlatform.Linux && UnixEnvironment.IsRoot()) {
+        //     usingEvDevInput = true;
+        //     InputManager.RegisterInputMethod(new EvDevKeyboardInputMethod());
+        // }
+        //
+        // if (this.WindowManager is SilkWindowManager silkWindowManager) {
+        //     if(!usingEvDevInput)
+        //         InputManager.RegisterInputMethod(InputManager.SilkWindowingKeyboardInputMethod = new SilkWindowingKeyboardInputMethod(silkWindowManager.InputContext));
+        //
+        //     InputManager.RegisterInputMethod(InputManager.SilkWindowingMouseInputMethod = new SilkWindowingMouseInputMethod(silkWindowManager.InputContext));
+        //     InputManager.SilkWindowingMouseInputMethod.SetRawInputStatus(FurballConfig.Instance.RawMouseInput);
+        // }
 
         Profiler.StartProfile("init_audio_engine");
         //TODO: Add logic to decide on what audio backend to use, and maybe write some code to help change backend on the fly
@@ -276,8 +276,8 @@ public class FurballGame : Game {
         MonoModCheck();
 
         if (RuntimeInfo.IsDebug()) {
-            this.CreateTextureDebugger();
-            this.CreateSceneViewer();
+            // this.CreateTextureDebugger();
+            // this.CreateSceneViewer();
         }
 
         WindowManager.FramebufferResize += this.OnWindowResize;
@@ -391,6 +391,8 @@ public class FurballGame : Game {
         catch { /* */ }
 
         this.UnregisterKeybinds();
+        
+        InputManager.End();
         
         GameTimeScheduler.Dispose(Time);
         EtoHelper.Dispose();
@@ -689,87 +691,87 @@ public class FurballGame : Game {
 
         DrawableBatch.End();
         
-        //We want to draw software cursors after everything else in the game, so we do it in here
-        #region Draw software cursors
-
-        List<FurballMouse> mice = InputManager.Mice;
-
-        bool needsRecalc = false;
-        foreach (FurballMouse furballMouse in mice) {
-            if (furballMouse.LastKnownSoftwareCursorPosition != furballMouse.Position)
-                needsRecalc = true;
-
-            furballMouse.LastKnownSoftwareCursorPosition = furballMouse.Position;
-        }
-
-        // ReSharper disable once CompareOfFloatsByEqualityOperator
-        if (this._cursorVerticalRatioCache != VerticalRatio)
-            needsRecalc = true;
-        
-        this._cursorVerticalRatioCache = VerticalRatio;
-        
-        if(needsRecalc) {
-            CursorDrawableBatch.Begin();
-
-            unsafe {
-                foreach (FurballMouse mouse in mice) {
-                    if (mouse.SoftwareCursor) {
-                        //3 vertices for the head, and 4 for the tail
-                        //3 indices for the head, 6 for the tail (2 tris)
-                        MappedData map = CursorDrawableBatch.Reserve(3 + 4, 3 + 6, WhitePixel);
-
-                        const float mouseWidth  = 15;
-                        const float mouseHeight = 25;
-                        const float tailWidth   = mouseWidth / 3f;
-
-                        float mouseScale = (float)(1 / VerticalRatio * FurballConfig.Instance.SoftwareCursorScale);
-
-                        map.VertexPtr[0].Position = Vector2.Zero * mouseScale + mouse.Position;
-                        map.VertexPtr[1].Position = new Vector2(0,          mouseHeight) * mouseScale + mouse.Position;
-                        map.VertexPtr[2].Position = new Vector2(mouseWidth, mouseHeight) * mouseScale + mouse.Position;
-
-                        const float tailLeft         = mouseWidth / 2f - tailWidth / 2f;
-                        const float tailRight        = mouseWidth / 2f + tailWidth / 2f;
-                        const float tailLength       = 10f;
-                        const float tailBottom       = mouseHeight + tailLength;
-                        const float tailBottomOffset = 2.5f;
-
-                        //Top left corner of the tail
-                        map.VertexPtr[3].Position = new Vector2(tailLeft, mouseHeight) * mouseScale + mouse.Position;
-                        //Top right corner of the tail
-                        map.VertexPtr[4].Position = new Vector2(tailRight, mouseHeight) * mouseScale + mouse.Position;
-                        //Bottom left corner of the tail
-                        map.VertexPtr[5].Position = new Vector2(tailLeft + tailBottomOffset, tailBottom) * mouseScale + mouse.Position;
-                        //Bottom right corner of the tail
-                        map.VertexPtr[6].Position = new Vector2(tailRight + tailBottomOffset, tailBottom) * mouseScale + mouse.Position;
-                        
-                        for (int i = 0; i < map.VertexCount; i++) {
-                            map.VertexPtr[i].Color = new Vixie.Backends.Shared.Color(1f, 1f, 1f, 0.8f);
-                            map.VertexPtr[i].TexId = map.TextureId;
-                        }
-                        
-                        //Head tri
-                        map.IndexPtr[0] = (ushort)(0 + map.IndexOffset);
-                        map.IndexPtr[1] = (ushort)(1 + map.IndexOffset);
-                        map.IndexPtr[2] = (ushort)(2 + map.IndexOffset);
-                        
-                        //Tail tri 1
-                        map.IndexPtr[3] = (ushort)(3 + map.IndexOffset);
-                        map.IndexPtr[4] = (ushort)(5 + map.IndexOffset);
-                        map.IndexPtr[5] = (ushort)(4 + map.IndexOffset);
-                        //Tail tri 2
-                        map.IndexPtr[6] = (ushort)(4 + map.IndexOffset);
-                        map.IndexPtr[7] = (ushort)(5 + map.IndexOffset);
-                        map.IndexPtr[8] = (ushort)(6 + map.IndexOffset);
-                    }
-                }
-            }
-
-            CursorDrawableBatch.SoftEnd();
-        }
-        CursorDrawableBatch.ManualDraw();
-
-        #endregion
+        // //We want to draw software cursors after everything else in the game, so we do it in here
+        // #region Draw software cursors
+        //
+        // List<FurballMouse> mice = InputManager.Mice;
+        //
+        // bool needsRecalc = false;
+        // foreach (FurballMouse furballMouse in mice) {
+        //     if (furballMouse.LastKnownSoftwareCursorPosition != furballMouse.Position)
+        //         needsRecalc = true;
+        //
+        //     furballMouse.LastKnownSoftwareCursorPosition = furballMouse.Position;
+        // }
+        //
+        // // ReSharper disable once CompareOfFloatsByEqualityOperator
+        // if (this._cursorVerticalRatioCache != VerticalRatio)
+        //     needsRecalc = true;
+        //
+        // this._cursorVerticalRatioCache = VerticalRatio;
+        //
+        // if(needsRecalc) {
+        //     CursorDrawableBatch.Begin();
+        //
+        //     unsafe {
+        //         foreach (FurballMouse mouse in mice) {
+        //             if (mouse.SoftwareCursor) {
+        //                 //3 vertices for the head, and 4 for the tail
+        //                 //3 indices for the head, 6 for the tail (2 tris)
+        //                 MappedData map = CursorDrawableBatch.Reserve(3 + 4, 3 + 6, WhitePixel);
+        //
+        //                 const float mouseWidth  = 15;
+        //                 const float mouseHeight = 25;
+        //                 const float tailWidth   = mouseWidth / 3f;
+        //
+        //                 float mouseScale = (float)(1 / VerticalRatio * FurballConfig.Instance.SoftwareCursorScale);
+        //
+        //                 map.VertexPtr[0].Position = Vector2.Zero * mouseScale + mouse.Position;
+        //                 map.VertexPtr[1].Position = new Vector2(0,          mouseHeight) * mouseScale + mouse.Position;
+        //                 map.VertexPtr[2].Position = new Vector2(mouseWidth, mouseHeight) * mouseScale + mouse.Position;
+        //
+        //                 const float tailLeft         = mouseWidth / 2f - tailWidth / 2f;
+        //                 const float tailRight        = mouseWidth / 2f + tailWidth / 2f;
+        //                 const float tailLength       = 10f;
+        //                 const float tailBottom       = mouseHeight + tailLength;
+        //                 const float tailBottomOffset = 2.5f;
+        //
+        //                 //Top left corner of the tail
+        //                 map.VertexPtr[3].Position = new Vector2(tailLeft, mouseHeight) * mouseScale + mouse.Position;
+        //                 //Top right corner of the tail
+        //                 map.VertexPtr[4].Position = new Vector2(tailRight, mouseHeight) * mouseScale + mouse.Position;
+        //                 //Bottom left corner of the tail
+        //                 map.VertexPtr[5].Position = new Vector2(tailLeft + tailBottomOffset, tailBottom) * mouseScale + mouse.Position;
+        //                 //Bottom right corner of the tail
+        //                 map.VertexPtr[6].Position = new Vector2(tailRight + tailBottomOffset, tailBottom) * mouseScale + mouse.Position;
+        //                 
+        //                 for (int i = 0; i < map.VertexCount; i++) {
+        //                     map.VertexPtr[i].Color = new Vixie.Backends.Shared.Color(1f, 1f, 1f, 0.8f);
+        //                     map.VertexPtr[i].TexId = map.TextureId;
+        //                 }
+        //                 
+        //                 //Head tri
+        //                 map.IndexPtr[0] = (ushort)(0 + map.IndexOffset);
+        //                 map.IndexPtr[1] = (ushort)(1 + map.IndexOffset);
+        //                 map.IndexPtr[2] = (ushort)(2 + map.IndexOffset);
+        //                 
+        //                 //Tail tri 1
+        //                 map.IndexPtr[3] = (ushort)(3 + map.IndexOffset);
+        //                 map.IndexPtr[4] = (ushort)(5 + map.IndexOffset);
+        //                 map.IndexPtr[5] = (ushort)(4 + map.IndexOffset);
+        //                 //Tail tri 2
+        //                 map.IndexPtr[6] = (ushort)(4 + map.IndexOffset);
+        //                 map.IndexPtr[7] = (ushort)(5 + map.IndexOffset);
+        //                 map.IndexPtr[8] = (ushort)(6 + map.IndexOffset);
+        //             }
+        //         }
+        //     }
+        //
+        //     CursorDrawableBatch.SoftEnd();
+        // }
+        // CursorDrawableBatch.ManualDraw();
+        //
+        // #endregion
 
         DrawCounts.LastVertexCount = Renderer.VertexCount;
         DrawCounts.LastIndexCount  = Renderer.IndexCount;
