+++ Furball.Game/FurballTestGame.cs
@@ -1,6 +1,6 @@
 ﻿using Furball.Engine;
-using Furball.Engine.Engine.Input;
 using Furball.Engine.Engine.Localization;
+using Furball.Engine.Engine.Input;
 using Furball.Game.Screens;
 using Furball.Vixie;
 using Silk.NET.Input;
